amend: rename asNodes to asChildNodes
