+++ packages/dataparcels-docs/src/docs/api/parcelHoc/modifyBeforeUpdate.md
@@ -10,7 +10,7 @@ type ParcelUpdater = (value: any, changeRequest: ChangeRequest) => any;
 
 // updating shape, including meta
 type ParcelUpdater = asNode((node: any, changeRequest: ChangeRequest) => any);
-type ParcelUpdater = asNodes((nodes: any, changeRequest: ChangeRequest) => any);
+type ParcelUpdater = asChildNodes((nodes: any, changeRequest: ChangeRequest) => any);
 type ParcelUpdater = asShape((parcelShape: ParcelShape, changeRequest: ChangeRequest) => any);
 ```
