+++ packages/dataparcels-docs/src/pages/api/useParcelForm.mdx
@@ -282,7 +282,7 @@ beforeChange?: ParcelUpdater|ParcelUpdater[] // optional
 
 type ParcelUpdater = (value: any, changeRequest: ChangeRequest) => any
 type ParcelUpdater = asNode((node: ParcelNode, changeRequest: ChangeRequest) => ParcelNode);
-type ParcelUpdater = asNodes((nodes: any, changeRequest: ChangeRequest) => any);
+type ParcelUpdater = asChildNodes((nodes: any, changeRequest: ChangeRequest) => any);
 type ParcelUpdater = asShape((parcelShape: ParcelShape, changeRequest: ChangeRequest) => any);
 ```
