+++ packages/dataparcels/__test__/Exports-test.js
@@ -10,7 +10,7 @@ import ParcelShape from '../ParcelShape';
 import ParcelNode from '../ParcelNode';
 import asShape from '../asShape';
 import asNode from '../asNode';
-import asNodes from '../asNodes';
+import asChildNodes from '../asChildNodes';
 import cancel from '../cancel';
 import validation from '../validation';
 
@@ -26,7 +26,7 @@ import InternalParcelShape from '../lib/parcelShape/ParcelShape';
 import InternalShape from '../lib/parcelShape/asShape';
 import InternalParcelNode from '../lib/parcelNode/ParcelNode';
 import InternalAsNode from '../lib/parcelNode/asNode';
-import InternalAsNodes from '../lib/parcelNode/asNodes';
+import InternalAsNodes from '../lib/parcelNode/asChildNodes';
 import Internalcancel from '../lib/change/cancel';
 import InternalValidation from '../lib/validation/validation';
 
@@ -66,8 +66,8 @@ test('/asNode should export asNode', () => {
     expect(asNode).toBe(InternalAsNode);
 });
 
-test('/asNodes should export asNodes', () => {
-    expect(asNodes).toBe(InternalAsNodes);
+test('/asChildNodes should export asChildNodes', () => {
+    expect(asChildNodes).toBe(InternalAsNodes);
 });
 
 test('/cancel should export cancel', () => {
