+++ packages/dataparcels/asChildNodes.js
@@ -1,0 +1,2 @@
+/* eslint-disable */
+module.exports = require('./lib/parcelNode/asChildNodes.js').default;
