--- packages/dataparcels/asNodes.js
@@ -1,2 +1,0 @@
-/* eslint-disable */
-module.exports = require('./lib/parcelNode/asNodes.js').default;
