+++ packages/dataparcels/package.json
@@ -18,7 +18,7 @@
     "ParcelNode.js",
     "ParcelShape.js",
     "asNode.js",
-    "asNodes.js",
+    "asChildNodes.js",
     "asRaw.js",
     "asShape.js",
     "validation.js"
