+++ packages/react-dataparcels/__test__/Exports-test.js
@@ -10,7 +10,7 @@ import ParcelShape from '../ParcelShape';
 import ParcelNode from '../ParcelNode';
 import asShape from '../asShape';
 import asNode from '../asNode';
-import asNodes from '../asNodes';
+import asChildNodes from '../asChildNodes';
 import cancel from '../cancel';
 import validation from '../validation';
 
@@ -34,7 +34,7 @@ import InternalParcelShape from 'dataparcels/ParcelShape';
 import InternalParcelNode from 'dataparcels/ParcelNode';
 import InternalShape from 'dataparcels/asShape';
 import InternalAsNode from 'dataparcels/asNode';
-import InternalAsNodes from 'dataparcels/asNodes';
+import InternalAsNodes from 'dataparcels/asChildNodes';
 import Internalcancel from 'dataparcels/cancel';
 import InternalValidation from 'dataparcels/validation';
 
@@ -84,8 +84,8 @@ test('/asNode should export asNode', () => {
     expect(asNode).toBe(InternalAsNode);
 });
 
-test('/asNodes should export asNodes', () => {
-    expect(asNodes).toBe(InternalAsNodes);
+test('/asChildNodes should export asChildNodes', () => {
+    expect(asChildNodes).toBe(InternalAsNodes);
 });
 
 test('/cancel should export cancel', () => {
