+++ packages/react-dataparcels/asChildNodes.js
@@ -1,0 +1,2 @@
+/* eslint-disable */
+module.exports = require('dataparcels/asChildNodes.js');
